requirements.txt
fastapi
uvicorn
pydantic
python-dotenv
langchain
langchain-community
langchain-core
langchain-google-genai
deep-translator
langdetect
